==> Semanas de trabajo/Semana 2/01_ETL_version/01_analisis_Energy_CO2.ipynb <==
import pandas as pd

## Proceso de limpieza dataset **energy**


df_emisionesCO2 = pd.read_csv('https://raw.githubusercontent.com/Sandtw/Proyecto-Grupal-EmisionesCO2/main/rawDataset/energyco2.csv')

df_emisionesCO2.head()

##### ***Eliminación de columnas no servibles***

df_emisionesCO2.drop('Unnamed: 0', axis = 1, inplace = True)

df_emisionesCO2.info()

##### ***Imputación de valores faltantes por 0***

df_emisionesCO2 = df_emisionesCO2.fillna(0)

df_emisionesCO2.info()

***Toma de datos, desde la fecha considerada (1980)***

df_emisionesCO2 = df_emisionesCO2[df_emisionesCO2['Year'] >= 1980]

##### ***Cambiar Quads a Terawatt por hora***

df_emisionesCO2['Energy_consumption'] = df_emisionesCO2['Energy_consumption'] * 293.07
df_emisionesCO2['Energy_production'] = df_emisionesCO2['Energy_production'] * 293.07

df_emisionesCO2 = df_emisionesCO2.round(3)
pd.set_option('display.float_format', lambda x: '%.3f' % x)

##### ***Normalizado de los valores de columna paises***

countries_list = ['Afghanistan', 'Albania', 'Algeria', 'American Samoa', 'Angola', 'Antarctica', 'Antigua and Barbuda', 'Argentina', 'Armenia', 'Aruba', 'Australia', 'Austria', 'Azerbaijan', 'Bahamas', 'Bahrain', 'Bangladesh', 'Barbados', 'Belarus', 'Belgium', 'Belize', 'Benin', 'Bermuda', 'Bhutan', 'Bolivia', 'Bosnia and Herzegovina', 'Botswana', 'Brazil', 'British Virgin Islands', 'Brunei', 'Bulgaria', 'Burkina Faso', 'Burundi', 'Cambodia', 'Cameroon', 'Canada', 'Cape Verde', 'Cayman Islands', 'Central African Republic', 'Chad', 'Chile', 'China', 'Colombia', 'Comoros', 'Congo', 'Cook Islands', 'Costa Rica', "Cote d'Ivoire", 'Croatia', 'Cuba', 'Cyprus', 'Czechia', 'Czechoslovakia', 'Democratic Republic of Congo', 'Denmark', 'Djibouti', 'Dominica', 'Dominican Republic', 'Ecuador', 'Egypt', 'El Salvador', 'Equatorial Guinea', 'Eritrea', 'Estonia', 'Eswatini', 'Ethiopia', 'Faeroe Islands', 'Falkland Islands', 'Fiji', 'Finland', 'France', 'French Guiana', 'French Polynesia', 'Gabon', 'Gambia', 'Georgia', 'Germany', 'Ghana', 'Gibraltar', 'Greece', 'Greenland', 'Grenada', 'Guadeloupe', 'Guam', 'Guatemala', 'Guinea', 'Guinea Bissau', 'Guyana', 'Haiti', 'Honduras', 'Hong Kong', 'Hungary', 'Iceland', 'India', 'Indonesia', 'Iran', 'Iraq', 'Ireland', 'Israel', 'Italy', 'Jamaica', 'Japan', 'Jordan', 'Kazakhstan', 'Kenya', 'Kiribati', 'Kosovo', 'Kuwait', 'Kyrgyzstan', 'Laos', 'Latvia', 'Lebanon', 'Lesotho', 'Liberia', 'Libya', 'Lithuania', 'Luxembourg', 'Macao', 'Madagascar', 'Malawi', 'Malaysia', 'Maldives', 'Mali', 'Malta', 'Martinique', 'Mauritania', 'Mauritius', 'Mexico', 'Micronesia', 'Moldova', 'Mongolia', 'Montenegro', 'Montserrat', 'Morocco', 'Mozambique', 'Myanmar', 'Namibia', 'Nauru', 'Nepal', 'Netherlands', 'Netherlands Antilles', 'New Caledonia', 'New Zealand', 'Nicaragua', 'Niger', 'Nigeria', 'Niue', 'North Korea', 'North Macedonia', 'Northern Mariana Islands', 'Norway', 'Oman', 'Pakistan', 'Palestine', 'Panama', 'Papua New Guinea', 'Paraguay', 'Peru', 'Philippines', 'Poland', 'Portugal', 'Puerto Rico', 'Qatar', 'Reunion', 'Romania', 'Russia', 'Rwanda', 'Saint Helena', 'Saint Kitts and Nevis', 'Saint Lucia', 'Saint Pierre and Miquelon', 'Saint Vincent and the Grenadines', 'Samoa', 'Sao Tome and Principe', 'Saudi Arabia', 'Senegal', 'Serbia', 'Seychelles', 'Sierra Leone', 'Singapore', 'Slovakia', 'Slovenia', 'Solomon Islands', 'Somalia', 'South Africa', 'South Korea', 'South Sudan', 'Spain', 'Sri Lanka', 'Sudan', 'Suriname', 'Sweden', 'Switzerland', 'Syria', 'Taiwan', 'Tajikistan', 'Tanzania', 'Thailand', 'Timor-Leste', 'Togo', 'Tonga', 'Trinidad and Tobago', 'Tunisia', 'Turkey', 'Turkmenistan', 'Turks and Caicos Islands', 'Tuvalu', 'Uganda', 'Ukraine', 'United Arab Emirates', 'United Kingdom', 'United States', 'United States Virgin Islands', 'Uruguay', 'Uzbekistan', 'Vanuatu', 'Venezuela', 'Vietnam', 'Western Sahara', 'Yemen',   'Yugoslavia', 'Zambia', 'Zimbabwe']

len(countries_list)

df_emisionesCO2 = df_emisionesCO2.replace('Cabo Verde','Cape Verde')
df_emisionesCO2 = df_emisionesCO2.replace('U.S. Virgin Islands','United States Virgin Islands')
df_emisionesCO2 = df_emisionesCO2.replace('Côte d’Ivoire',"Cote d'Ivoire")
df_emisionesCO2 = df_emisionesCO2.replace('Guinea-Bissau','Gambia')
df_emisionesCO2 = df_emisionesCO2.replace('Czech Republic','Czechia')
df_emisionesCO2 = df_emisionesCO2.replace('The Bahamas','Bahamas')
df_emisionesCO2 = df_emisionesCO2.replace('Burma','Myanmar')
df_emisionesCO2 = df_emisionesCO2.replace('Macau','Macao')
df_emisionesCO2 = df_emisionesCO2.replace('Gambia, The','Gambia')

df_emisionesCO2_by_country = df_emisionesCO2[df_emisionesCO2.Country.isin(countries_list)]

##### ***Registro de agrupaciones de países***

df_emisionesCO2_world = df_emisionesCO2[df_emisionesCO2['Country'] == 'World']

df_emisionesCO2_groups = df_emisionesCO2[~df_emisionesCO2.Country.isin(countries_list)]
df_emisionesCO2_groups['Country'].unique()

#### ***Normalización de columnas***

columns_norm = ["country", "source", 'year', 'consumptionElectricity', 'productionElectricity', 'gdp', 'population', 'intensityPerCapita', 'intensityByGDP', 'co2emissions']

df_emisionesCO2_by_country.columns = columns_norm

df_emisionesCO2_by_country.head()

Normalización de tabla ***'df_emisionesCO2_by_country'***

Considerando las fuentes de energia sin agrupacion

co2_emissions = df_emisionesCO2_by_country

co2_emissions.head()

##### ***Tabla source_energy***

import numpy as np
sources = np.sort(np.array(['all_energy_types', 'biomass', 'coal', 'cogeneration', 'gas','geothermal', 'hydro', 'nuclear', 'oil', 'other', 'other_renewables', 'petcoke', 'renewables', 'solar', 'storage', 'waste', 'wave and tidal', 'wind']))

source_energy = pd.DataFrame(sources.reshape(-1,1), columns = ['source']).reset_index(drop=True)

source_energy = source_energy.reset_index().rename(columns={'index':'idSource'})
source_energy

source_energy.to_csv('../Modelo_ER/source_energy.csv', index=False)

##### ***Tabla emission***

Consistencia de valores de las fuentes de energia

- renewables_n_other : other_renewables
- petroleum_n_other_liquids: oil
- natural_gas:gas

co2_emissions = co2_emissions.replace('renewables_n_other','other_renewables')
co2_emissions = co2_emissions.replace('petroleum_n_other_liquids','oil')
co2_emissions = co2_emissions.replace('natural_gas','gas')



columns_emission = ['country', 'year','source','co2emissions']

emission = co2_emissions[columns_emission]

emission.drop_duplicates(subset = ['country', 'year','source'], inplace=True)

emission.head()

emission.reset_index(drop=True).to_csv('../Modelo_ER/emission.csv', index=False)

##### ***Tabla source_country_electricity de energyCO2***

source_country_electricity_energyCO2 = co2_emissions[['country', 'year', 'source', 'consumptionElectricity', 'productionElectricity']].sort_values(by=['country','year', 'source'])

source_country_electricity_energyCO2 = source_country_electricity_energyCO2[source_country_electricity_energyCO2['source'] != 'all_energy_types']

source_country_electricity_energyCO2.head()

source_country_electricity_energyCO2.drop_duplicates(subset = ['country','year','source'], inplace = True)

source_country_electricity_energyCO2.sort_values(by=['country','year','source']).reset_index(drop=True).to_csv('../normalizedDataset_2/source_country_electricity_energyCO2.csv', index=False)

##### **Subida a github**

from github import Github
token = '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'
g = Github(token)
repo = g.get_repo('Sandtw/Proyecto-Grupal-EmisionesCO2')

try:
    repo.create_file('energyco2_by_country.csv','Limpieza emisiones CO2 por país', df_emisionesCO2_by_country.to_csv( index = False))
except:
    contents = repo.get_contents('energyco2_by_country.csv')
    repo.update_file(contents.path, 'Actualización, limpieza CO2 por país', df_emisionesCO2_by_country.to_csv(index = False), contents.sha)

try:
    repo.create_file('energyco2_worlds.csv','Limpieza emisiones C02 mundial', df_emisionesCO2_world.to_csv( index = False))
except:
    contents = repo.get_contents('energyco2_worlds.csv')
    repo.update_file(contents.path, 'Actualización, limpieza CO2 mundial', df_emisionesCO2_world.to_csv(index = False), contents.sha)
