==> Semanas de trabajo/Semana 2/01_ETL_version/05_tables_keys.ipynb <==
import numpy as np
import pandas as pd

country = pd.read_csv('../Modelo_ER/country.csv')
emission = pd.read_csv('../Modelo_ER/emission.csv')
plant = pd.read_csv('../Modelo_ER/plant.csv')
population_country_electricity = pd.read_csv('../Modelo_ER/population_country_electricity.csv')
source_country_electricity = pd.read_csv('../Modelo_ER/source_country_electricity.csv')
source_energy = pd.read_csv('../Modelo_ER/source_energy.csv')
region = pd.read_csv('../Modelo_ER/region.csv')
emission_region = pd.read_csv('../Modelo_ER/emission_region.csv')
source_region_electricity = pd.read_csv('../Modelo_ER/source_region_electricity.csv')

### Acoplando clave en emission (idCountry, idSource)

emission.insert(0, 'idCountry', '')
emission.insert(3, 'idSource', '')

# idCountry
emission = emission.merge(country, on = ['country'])
emission.head()

emission['idCountry_x'] = emission['idCountry_y']
emission = emission.rename(columns = {'idCountry_x':'idCountry'})
emission.drop(['idCountry_y', 'iso_code', 'country'], axis = 1, inplace=True)


# idSource
emission = emission.merge(source_energy, on = ['source'])
emission.head()

emission['idSource_x'] = emission['idSource_y']
emission = emission.rename(columns = {'idSource_x':'idSource'})
emission.drop(['idSource_y', 'source'], axis = 1, inplace=True)

emission.to_csv('../Modelo_ER_2/emission.csv', index=False)

### Acoplando clave en plant (idCountry)

plant.insert(0, 'idCountry', '')
plant.insert(1, 'idPrimaryFuel', '')

# idCountry
plant = plant.merge(country, on = ['country'])
plant.head()

plant['idCountry_x'] = plant['idCountry_y']
plant = plant.rename(columns = {'idCountry_x':'idCountry'})
plant.drop(['idCountry_y', 'iso_code', 'country'], axis = 1, inplace=True)

# idPrimaryFuel
plant = plant.merge(source_energy, left_on = ['primaryfuel'], right_on = ['source'])
plant.head()

plant['idPrimaryFuel'] = plant['idSource']
plant.drop(['idSource', 'source', 'primaryfuel'], axis = 1, inplace=True)

plant

plant.sort_values(['idCountry','namePlant']).to_csv('../Modelo_ER_2/plant.csv', index=False)

### Acoplando clave en population_country_electricity (idCountry)

population_country_electricity.insert(0, 'idCountry', '')

population_country_electricity

# idCountry
population_country_electricity = population_country_electricity.merge(country, on = ['country'])
population_country_electricity.head()

population_country_electricity['idCountry_x'] = population_country_electricity['idCountry_y']
population_country_electricity = population_country_electricity.rename(columns = {'idCountry_x':'idCountry'})
population_country_electricity.drop(['idCountry_y', 'iso_code', 'country'], axis = 1, inplace=True)

population_country_electricity.head()

population_country_electricity.sort_values(['idCountry','year']).to_csv('../Modelo_ER_2/population_country_electricity.csv', index=False)

### Acoplando clave en source_country_electricity (idCountry, id_source)

source_country_electricity.insert(0, 'idCountry', '')
source_country_electricity.insert(3, 'idSource', '')


# idCountry
source_country_electricity = source_country_electricity.merge(country, on = ['country'])
source_country_electricity.head()

source_country_electricity['idCountry_x'] = source_country_electricity['idCountry_y']
source_country_electricity = source_country_electricity.rename(columns = {'idCountry_x':'idCountry'})
source_country_electricity.drop(['idCountry_y', 'iso_code', 'country'], axis = 1, inplace=True)

source_country_electricity

# idSource
source_country_electricity = source_country_electricity.merge(source_energy, on = ['source'])
source_country_electricity.head()

source_country_electricity['idSource_x'] = source_country_electricity['idSource_y']
source_country_electricity = source_country_electricity.rename(columns = {'idSource_x':'idSource'})
source_country_electricity.drop(['idSource_y', 'source'], axis = 1, inplace=True)

source_country_electricity

source_country_electricity.to_csv('../Modelo_ER_2/source_country_electricity.csv', index=False)

### Acoplando clave en emission_region (idRegion)

emission_region.insert(0, 'idRegion', '')

# idCountry
emission_region = emission_region.merge(region, on = ['region'])
emission_region.head()

emission_region['idRegion_x'] = emission_region['idRegion_y']
emission_region = emission_region.rename(columns = {'idRegion_x':'idRegion'})
emission_region.drop(['idRegion_y','region'], axis = 1, inplace=True)

emission_region.sort_values(['idRegion','year']).to_csv('../Modelo_ER_2/emission_region.csv', index=False)

### Acoplando clave en source_region_electricity (idRegion,idSource)

source_region_electricity.insert(0, 'idRegion', '')
source_region_electricity.insert(3, 'idSource', '')

# idRegion
source_region_electricity = source_region_electricity.merge(region, on = ['region'])
source_region_electricity.head()

source_region_electricity['idRegion_x'] = source_region_electricity['idRegion_y']
source_region_electricity = source_region_electricity.rename(columns = {'idRegion_x':'idRegion'})
source_region_electricity.drop(['idRegion_y', 'region'], axis = 1, inplace=True)

source_region_electricity

# idSource
source_region_electricity = source_region_electricity.merge(source_energy, on = ['source'])
source_region_electricity.head()

source_region_electricity['idSource_x'] = source_region_electricity['idSource_y']
source_region_electricity = source_region_electricity.rename(columns = {'idSource_x':'idSource'})
source_region_electricity.drop(['idSource_y', 'source'], axis = 1, inplace=True)

source_region_electricity

source_region_electricity.to_csv('../Modelo_ER_2/source_region_electricity.csv', index=False)
